refactor pasapi server.go
